EmptyDataView properly implemented
This adds the empty data view to each of the main view of the application, to provide information in how to alleviate the situation. Along with the messages buttons have been added to allow users to quickly attempt to resolve the issue. Future plan is to auto navigate to the view that will help solve the issue.

- BrowseController.kt       : Using string res & action for EmptyDataView.kt, Inlined refreshExtensions into onRefresh
- DownloadsController.kt    : Added showEmpty override
- EmptyDataView.kt          : Added new vararg show overload
- LibraryController.kt      : Using string res for EmptyDataView.kt
- RepositoryController.kt   : Added string res & action for EmptyDataView.kt, separated launchAddRepositoryDialog
- UpdatesController.kt      : Added showEmpty override
- RepositoryUpdateWorker.kt : Fixed deprecation
- strings.xml               : new strings for EmptyDataView.kt