Add EmptyDataView & ProgressBar (settings) in ConfigureExtension.kt
This is a mostly cosmetic update.
It informs the user that there can be settings for an Extension.
For slower devices, Progress indicator for the loading of the settings.

Revealed an issue that on first boot, the ViewModel wont emit empty.
This is fixed with an auto check after 200ms to reload.