Implement new update technique
Previous update style, while faster, did not account for various
situations that now show up with multiple repositories.

The entire update equation is now handled by the RepositoryUpdateWorker,
In the future it would be a good idea to move it into the SOT, but this
can be ignored for now.