Implement External Backup Restoring
Inorder to achieve this, I had to hack around with how Android provides
a file Uri from the file selection. It seems the Uri object itself is
important to the loading process, thus I needed to pass the object to
the RestoreBackupWorker.

This release also adds a progress to the restore chapters notification
:D.