Normalization & Purge of HResult
God I hate this with a passion.

HResult created a mess in the code base that I am now unraveling.
Normalization of the database is underway. Currently I want remove
 HResult from the codebase. Upon completion the code base should be less
 prone to crashes due to properly handling exceptions.

Asides from that normalization of extensions should not be too hard.

This is on a separate branch so I can keep developing development while
 i work on this partial rewrite.