Set MutableStateFlow.value instead of .emit
MutableStateFlow is made for watching a state of a value, if you set .value it will first check if its equal to the last value and only emit the new value if its not. You shouldn't use .emit or .tryEmit on a MutableStateFlow, if you need a flow where you can emit independent of the previous value, while keeping track of the last value, you should be using a MutableSharedFlow with a replay value of 1.